+++ src/nvim/api/vim.c
@@ -52,6 +52,7 @@
 #include "nvim/msgpack_rpc/unpacker.h"
 #include "nvim/ops.h"
 #include "nvim/option.h"
+#include "nvim/optionstr.h"
 #include "nvim/os/input.h"
 #include "nvim/os/process.h"
 #include "nvim/popupmenu.h"
