+++ src/nvim/optionstr.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_OPTIONSTR_H
+#define NVIM_OPTIONSTR_H
+
+#include "nvim/buffer_defs.h"  // for buf_T, win_T
+#include "nvim/option_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "optionstr.h.generated.h"
+#endif
+#endif  // NVIM_OPTIONSTR_H
